Use existing exchangeCommandHandler as wireGuardHandler for platforms which do not support wgctrl-go
